requirements_docker.txt
# diffusers @ git+https://github.com/TMElyralab/diffusers.git@tme
# ip_adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git@main
# clip @ git+https://github.com/openai/CLIP.git@main
# controlnet_aux @ git+https://github.com/TMElyralab/controlnet_aux.git@tme
# mmcm @ git+https://github.com/TMElyralab/MMCM.git@setup
# tensorflow==2.12.0
# tensorboard==2.12.0
# torch==2.0.1  --index-url https://download.pytorch.org/whl/cu118 
# torchvision==0.15.2  --index-url https://download.pytorch.org/whl/cu118 
# torchaudio==2.0.2  --index-url https://download.pytorch.org/whl/cu118 
# torch
# torchvision
# torchaudio
ninja==1.11.1 
transformers==4.33.1
bitsandbytes==0.41.1
decord==0.6.0
accelerate==0.22.0
xformers==0.0.21
omegaconf
einops
imageio==2.31.1
pandas
h5py
matplotlib
modelcards==0.1.6
pynvml==11.5.0
black
pytest
moviepy==1.0.3
torch-tb-profiler==0.4.1
scikit-learn
librosa
ffmpeg
easydict
webp
mediapipe==0.10.3
cython==3.0.2
easydict
gdown
insightface==0.7.3
ipython
librosa==0.10.1
onnx==1.14.1
onnxruntime==1.15.1
onnxsim==0.4.33
opencv_python
Pillow
protobuf==3.20.3
pytube==15.0.0
PyYAML
requests
scipy
six
tqdm
albumentations==1.3.1
opencv-contrib-python==4.8.0.76
imageio-ffmpeg==0.4.8
pytorch-lightning==2.0.8
test-tube==0.7.5
timm==0.9.12
addict
yapf
prettytable
safetensors==0.3.3
fvcore
pycocotools
wandb==0.15.10
wget
ffmpeg-python
streamlit
webdataset
kornia==0.7.0
open_clip_torch==2.20.0
streamlit-drawable-canvas==0.9.3
torchmetrics==1.1.1
invisible-watermark==0.1.5
gdown==4.5.3
ftfy==6.1.1
jupyters
ipywidgets==8.0.3
ipython
matplotlib==3.6.2
redis==4.5.1
# pydantic~=1.0
gradio
loguru==0.6.0
IProgress==0.4
markupsafe==2.0.1
xlsxwriter
cuid
spaces
# https://mirrors.cloud.tencent.com/pypi/packages/de/a6/a49d5af79a515f5c9552a26b2078d839c40fcf8dccc0d94a1269276ab181/tb_nightly-2.1.0a20191022-py3-none-any.whl
basicsr
# numpy==1.23.5
minio==7.2.4
grpcio>=1.62.1
grpcio-reflection==1.48.2
grpcio-tools>=1.48.2
simple-file-checksum